Editor screen got few improvements;  Server request utf-8 fix; Added requests for like;

+++ PopPicture/app/src/main/java/leonardolana/poppicture/server/ServerRequestRemovePicture.java
@@ -25,7 +25,7 @@ import leonardolana.poppicture.helpers.api.UserHelper;
  */
 public class ServerRequestRemovePicture extends ServerRequest {
 
-    public ServerRequestRemovePicture(int pictureId) {
+    public ServerRequestRemovePicture(long pictureId) {
         super(ServerConstants.URL, "Picture.remove");
         addParam(KEY_PICTURE_ID, pictureId);
     }
